[cv][cl] extra generic part of CL deps resolution from gerrit/updater.
This is a rewrite with a different trade off of the current code
https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/cv/internal/gerrit/updater/fetcher.go;l=652;drc=259a85dd70ad602f77dd82d93385dc432ffd03cd

For CLs with just 1 dep, this won't make a difference.
But for CLs with many deps, in ideal case, we'll have just two
batch DS.Get RPCs instead of 2*N before.

Unfortunately, this makes code more complicated, so I added a helper
struct (resolvingDep) to make higher level code clearer (or so I hope).

The biggest con of the new trade off is the likely exhaustion of RAM
if a CL has many deps. I've added a warning to help detect this
situation in production. If this happens, it'll be very easy to fix
as I have kept lifetime of a loaded (dep) CL entity small and contained
in the resolveDeps() helper function.

Next, there is a trade off in how to expose the new functionality
API/layer-wise. The requirements are:
 (1) updaterBackend.Fetch() returns a Snapshot with .Deps field,
     -> so CL entities for each dep must be created before that;
 (2) before the CL entity is saved with a new Snapshot{Deps:...},
     a UpdaterCLTask has to be scheduled for each dep which needs it
     (e.g. because CV didn't know about the dep before);
 (3) scheduling CL update tasks is done via Updater;
 (4) creating CL entities requires CL Mutator to facilitate
     notifications to other CV components.

I considered several approaches:
 (A) expose as Updater.ResolveAndScheduleDepsUpdate and let
     updaterBackend depend on the Updater
       - this feels circular, but I know this pattern is common
         (and I bet Java folks have a name for it, so I guess I won't be
         fired for it :)
       - a bit harder to test a specific updaterBackend

 (B) provide functionality as a callback to updaterBackend.Fetch()
       - essentially same as (A), but dependency is injected in a more
       obscure way,
       + a bit easier to test updaterBackend

 (C) change interface of Fetch s.t. it doesn't have to set
     Snapshot.Deps itself, instead returning additional value of
         map[ExternalID]DepKind
     which is then resolved by the Updater itself.
       + cleanest separation of concerns semantically,
       - but because .Deps are part of the Snapshot already, creates
         cognitive overhead and confusion that Snapshot returned by
         backend isn't actually written as is by the Updater.
         Notably, this complicates code of UpdateFields struct
         and Updater itself.
         -> can be mitigated by introducing a new type for Snapshot w/o
         resolved Deps and using it inside UpdateFields. However, since
         Snapshot is itself a proto, we can't split it backward
         compative way. Or, it can be a new type Go-native type, but then
         it has to copy all but 1 field from the proto, which is also
         not ideal.

I chose (A) so far as the lesser evil. But feedback is very welcome.
The good news is that adopting (B) or even (C) as a follow up is
really easy.

R=yiwzhang

Bug: 1281989
Change-Id: I6a81e6c23f0461fda5e8e4560e27090321e69854
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3352335
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>