[cv][cl] implement pessimistic 1 CL update.
The optimistic avoidance of fetch is left to the next CL,
as this CL is already huge.

This is a Gerrit-agnostic rewrite of the existing code,
but you don't have to read it to review the CL -- the new code
must be self-explanatory -- else please LMK!
Still, here is the reference to old code as FYI:
https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/cv/internal/gerrit/updater/fetcher.go;l=93;drc=259a85dd70ad602f77dd82d93385dc432ffd03cd

All tests added are high level to make them easier to understand
and test actual business requirements.

I've also split tests into separate TestXXX funcs as I found
edit-test-fail-fix cycle faster than hand-curating 4-levels of
"FocusConvey" blocks.

R=yiwzhang

Bug: 1281989
Change-Id: I442bbec0d26d6447545cf8c883ca2cb7d2b3c922
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3352336
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>