[cv] migrate to generic changelist.Updater.
Depends on http://cl/418972252 deployed first,
to ensure we don't lose alerting coverage.

**Danger**: this must be deployed carefully,
but rollback is safe.

That said, the diff is almost entirely "mechanical"
and the e2e/ tests didn't require any modifications,
so I have very high confidence that this is fine.

For backwards compatibility, keep the old gerrit/updater.Updater serving
existing tasks. It'll be deleted only if this CL sticks in production.

And if it doesn't, rollback is safe the code to handle the new kind of
CL update tasks has landed before this CL and should continue working
after the revert/rollback. But even if the new changelist.Updater
doesn't work after the rollback, it's OK to purge the task queue and
let Gerrit pollers and Run Managers re-create CL refresh tasks of the
old kind -- the outage should be fully resolved in ~30 minutes.

R=yiwzhang

Bug: 1281989
Change-Id: If16d969714b34f6571551168cfcb61ff6c9bd25a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3352340
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>