[verified-builds]: wrapper command proto addition
For task isolation purposes, we need a way to
"sandbox" the task that buildbucket runs. To implement
this, we introduce a "Wrapper" that will receive the "exe"
specified for the build as a parameter. The owner of the build
is responsible for making sure that the wrapper command exists
and is in the correct path. If the wrapper command does not exist,
the build will fail with INFRA_FAILURE.

For Linux, the "wrapper"command will likely be nsjail, or
something similar.

Implementing the sandboxing functionality in this way
provides flexibility in how the sandboxing/isolation technology
works across operating systems, and does not create an overly-specific
API that requires that buildbucket have deeper knowledge of
isolation primitives that are used.


Bug:1094754
Change-Id: I1a053cceaa7bbf16ed506e8c8a921c78b52128e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3309780
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>