[realms] Add minimal validation of Realms proto.
This is a light sanity check so if Realms proto is invalid for
whatever reason, we crash sooner than later.

The only way it can be invalid is when permissions and conditions
referenced in Binding by index are outside of defined permissions
and conditions lists.

R=iannucci@chromium.org
BUG=1280442

Change-Id: Id2e26c35d2864fedbc856088723fb67afc6b6b51
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3402724
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>