[tokenserver] Reject configs which share accounts between teams.
Now that every project in projects.cfg has an owning team (enforced by
the config service), we can make this validation better.

Previously this only issued an error when an account was shared between
projects, but this had to be softened to support chrome's intentional
sharing of an account between main and release branch projects.

Now that we have "owning team" information we can make it a hard error
to share an account between projects which belong to different teams. I
can't think of a reason why any team would want to do such a thing (or
should do such a thing), and no teams currently do this, so I figure
this is a good time to turn it back on.

Making it a blocking error should also prevent this case from occurring
in the future.

R=vadimsh

Change-Id: I640fb501a4b023ab51c34869316b705d96bbf12f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3408647
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>