[cv] update CheckResult to store per-cl failure message
There doesn't seem to be reasons to post a summary of
all failures to all CLs. If Run creatin fails, a message will be
posted into each CL for the reason of the CL failure.

This CL keeps summary for logging purpose, and it will be
removed later.

Bug: 1268574
Change-Id: I002055b786d618ba1c5ea281acbb7378fe973298
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3427572
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>