[auth-service]: GetRelevantSubgraph
GetRelevantSubgraph returns a Subgraph groups that include the principal and are
owned by the principal. Subgraph may include an extra node(s) for globs
or identity if the principal is a glob or identity.

This will be used for lookup when going to https://chrome-infra-auth.appspot.com/auth/lookup


Bug:1112832
Change-Id: I5b4b627b4b009654e6456aec1f05ebc729622650
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3384636
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Camilo Jacomet <cjacomet@google.com>