[buildbucket] Add configuration option to explicitly permit overriding properties.
Not yet enforced; The plan is to:
1. Land and deploy this change
2. Propose CLs to all Builders which currently accept ScheduleBuild requests with
   property overrides which intersect with the Builder's defined properties.
3. Start enforcing this setting.

The intent is to make it so that Builders which allow users to overwrite properties
explicitly indicate which properties are allowed to be overridden.

This will allow the definition of secure builders which don't permit sensitive
properties like 'recipe' to be overwritten, for potential exploit.

R=vadimsh

Bug: 1306248
Change-Id: I20b5018cc0bc77a1f413592ceaecb3f51b37a1c9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3537445
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>