[deploy] Remember last successfully applied intended state.
Will be used to trigger actuation when intended state changes,
even if the reported state still matches the intended state.

One case where this is useful is changes to `vars` section in
Starlark GAE configs: `vars` aren't reflected in `reported_state`
in any way. If we just use `reported_state` and compared it to the
current `inteded_state` we won't know that `vars` has changes. But
comparing the current `inteded_state` to the last applied intended
state would reveal that `vars` are different and we need to do
redeploy.

(In practice this will not actually work because infradata/gae
Makefiles don't handle this edge case either and will just skip
the deployment as well, but it is a separate issue).

R=iannucci@chromium.org, ddoman@chromium.org
BUG=1301023

Change-Id: I9b33103227140fd5fb77e7b6e4945bc64cd55090
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3566504
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>