[deploy] Handle multi-asset actuations better.
An actuation can theoretically "touch" several assets. Some of them
might be up-to-date. When this happens, such up-to-date assets are
still associated with the actuation via `LastActuation` field. This
field holds a full snapshot of Actuation, including its `State`
field.

Such up-to-date assets are not reported in EndActuation (because
they aren't really touched). But when the actuation finishes (goes
into e.g. SUCCEEDED state), its state (and other properties) need to
be updated in all assets that refer to it via `LastActuation` field,
even ones that were reported as up-to-date. Otherwise their
`LastActuation` field has stale Actuation.

R=ddoman@chromium.org, iannucci@chromium.org
BUG=1301023

Change-Id: I8d929d1b1e5dc18f4d2456428e9bf3ba3edbf38a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3582411
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>