[deploy] Add ListAssetHistory RPC.
It returns all information that is needed by the asset web UI page:
  * The current state of the asset.
  * The current active actuation (if any).
  * History of actuations.

Note that to get a consistent snapshot of the history we need to
fetch Asset and its currently recorded HistoryEntry transactionally
(older finalized history records are immutable and can be fetched
non-transactionally). That's the reason ListAssetHistory returns
Asset as well, instead of just returning AssetHistory records.

This is important for UI that shows the current and historical states
on the same page.

R=ddoman@chromium.org, iannucci@chromium.org
BUG=1301023

Change-Id: Ie3890877d4edeef3fb40fe93caf796ede936569b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3605841
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>