[deploy] Add a field with commit details to Deployment.
They are useful in the UI when showing the history of actuations.

They can theoretically be fetched dynamically by the backend, but
so far it didn't need direct access to the configs repo and I'd
like to keep it this way. It is arguably simpler to just grab it
on the builder side in the recipe.

R=ddoman@chromium.org, iannucci@chromium.org
BUG=1301023

Change-Id: If4561c0799a5f3468ed139767adfa33f5098ded2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3606386
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>