[ResultDB] Add exoneration reason field to ResultDB and ResultSink.
Exoneration reason is a machine-readable reason for the exoneration.
It will be used by Weetbix to avoid feeding the output of its own
analysis (or analyses it is replacing, such as FindIt) back as an
input.

Initially, only exonerations with the reason OCCURS_ON_OTHER_CLS
and OCCURS_ON_MAINLINE will be excluded by Weetbix. However, machine
readable versions of all exoneration explanations observed on a sample
of (1/256th) of data in prod have been generated.

These broadly cover the spectrum of reasons possible from:
- exonerated because not a failure, to
- exonerated because of insufficient test track record
  (e.g. informational, experimental tests), to
- exonerated because test track record shows similar failures
  occurring without patch or in other unrelated CLs (with patch).

In a follow-up CL, I will propose to perform a ParitionedUpdate
which sets a Reason of 0 for the Reason column and then change
the schema to mandate this column be not NULL.
This will allow expedited deprecation of the code path which
currently handles NULLs, and the associated tests.

BUG=1314194
TEST=go test go.chromium.org/luci/resultdb/...

Change-Id: Ia5ee67b39dfdbd7c39ee73e6cecab35407c08bfc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3607770
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>