cv: refactor tryjobs update handling and mark execution as succeeded if needed
I rewrote most of the logic to make things look cleaner. I did try to
add to the existing code but then I figured the code is no longer
readable. Also, this change fixed a bug that a failed tryjob that
is not included in the updated tryjobs list will be retried but the
code can't find its correspoding tryjob entity because it was never
loaded at the beginning. This bug was actually introduced due to one
of my previous bad suggestions in one ealier CL.

R=robertocn

Bug: 1312255
Change-Id: I8b7e734bd2daf509c3a05e3e05092866bd0927e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3633617
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>