[cipd] Log GS download speed when verifying packages.
Sometimes for no apparent reason downloads of particular packages
during hash verification are super slow. When this happens, one
potential solution is to abort and retry on another GAE instance.

Add some logging to understand what expected download speed should
be and what it is when it is "super slow" (and what packages exactly
are "super slow" to verify).

R=iannucci@chromium.org, ddoman@chromium.org
BUG=1261988

Change-Id: I8fcf078c45341a3781a52508b22fe51145b9c127
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3642568
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>