[gae] Make go.chromium.org/luci/gae work on go116 GAE.
It is a horrible hack. "google.golang.org/appengine" appears to be
mostly working on go116 GAE, except logging is broken. But luci/gae
wraps logging and we can implement it using GAE v2 method (structured
logging to stdout).

The only wrinkle in this plan is that the appengine library still
calls logservice.Flush RPC, even if there's no logs to flush (and
this fails and emits noise in stderr).

But there's an unreleased (!) 1.5y old commit in
github.com/golang/appengine that provides an undocumented way to
disable logging done by the library. We rely on it now by fetching
appengine module via git hash instead of a released version number.

To use this hack, the service should have the following in app.yaml:

  runtime: go116
  app_engine_apis: true

Q: Why not switch to google.golang.org/appengine/v2 (which
essentially does the same thing as this commit, just a layer
lower)?

A: To ease the migration. Due to how Go modules work, v2 is
a completely separate package. Unless we switch all code base (and
all dependencies) to v2 in a single commit, there'll be a possibility
of both v1 and v2 libraries linked into a single binary. Since both
of them register the same protobuf messages, it causes panics during
startup due to name conflicts in the proto registry:
https://github.com/golang/appengine/issues/281 All services will have
to be switch to go116 at once, otherwise they'll likely panic during
startup. With this hack, the exact same codebase works on both go111
and go116 and there are no panics. Individual services can switch to
go116 at their own pace after core libraries have switched.

Q: So once all go111 apps switch to go116 using this hack, we are
done with GAE migrations and can live happily ever after?

A: No, this is a fragile hack and a dead end. Code that relies on it
will not be getting any real support (i.e. new features) and very
likely will break at some point. LUCI services should still migrate
to Google Cloud APIs and go.chromium.org/luci/server framework.

Q: If it is such a horrible hack, and there's a better approach, why
are we doing it?

A: It is becoming extremely difficult to maintain buildability of
shared luci-go code on go111 for GAE. Many fundamental libraries
(like gRPC) deprecated go111 support loooong time ago, and we have
to resort to pinning dependencies to prehistoric versions. With more
and more libraries removing go111 compatibility, it is getting harder
and harder to maintain pinned dependencies. At this point it is less
frustrating to maintain an abominable "appengine v1 library on top
of go116 runtime with hacks" than try to keep go111 compatibility.

R=iannucci@chromium.org

Change-Id: Id2a8daf85d30efc4b82f4d177b1639b5a1d1894a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3651087
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>