Fix some misuse of the format string in status.Errorf.
The second argument in status.Errorf is a format string
(with "%s" and stuff). It is wrong to pass arbitrary runtime
string to it. If Go was C, it would have been a severe security
vulnerability, but in Go it just results in weird looking error
messages like:

  no progress made due to rpc error: code = Unknown
  desc = Get https://.../a/project%!F(MISSING)path%!F(MISSING).../+log/...

R=yiwzhang@google.com

Change-Id: Ib3bede7ff9f45bfa756f4604e14d6df3e3662cbb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3656174
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>