[machine-db] Switch to go116 runtime.
This is a noop essentially since machine-db doesn't exist anymore
and this code is not deployed anywhere.

I attempted to remove it, but there are references to machine-db
packages from "infra/unifiedfleet", so it is not as simple as
"rm -rf" and should be done separately in b/189864019.

R=iannucci@chromium.org
CC=anushruth@google.com

Change-Id: Iaf1294c0d923e42855a8fee82d5ef9f77ccf6a07
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3656371
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>