[lucicfg] Fix `unused-variable` lint in lucicfg code.
It is a new lint pulled in [1] and it appears to be generally
useful and not very buggy, so keep it enabled in the default set.

The only problem is that it is noisy in files dedicated for
generating docs (they have no implementation, all variables are
unused).

[1]: https://chromium.googlesource.com/infra/luci/luci-go/+/5bc92529c

R=iannucci@chromium.org, yiwzhang@google.com

Change-Id: Iffb34937581827274785ed7a5581b7d307841e33
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3665880
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>