[milo] add documentation to ACL config
This serves as a reminder on why we need to have this config.
The choice made in the design doc, however, may no longer be the best
choice given the change of teck stack. But the link to the doc is still
valuable when we eventually decide to migrate to a different solution.

R=mdraz@google.com, nqmtuan@google.com

Change-Id: I608eb0ff1332e910745fb8401a2182339e4d2b16
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3670659
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>