[cipd] Add `export` subcommand and $OverrideInstallMode.
`$OverrideInstallMode copy` allows an ensure file to specify that
everything it points to should be installed in copy-mode, even if
individual packages request to be installed with system-default or
symlink mode. As mentioned in the bug, symlink has a variety of
subtle issues which can make it difficult to use beyond just simple
"read file data" or "execute program" use cases.

The export subcommand lets you plop all the data from an ensure-file on
to disk, but without any of the internal '.cipd' metadata. This implies
setting `$OverrideInstallMode copy`. The removal of the metadata
directory  means that you cannot re-use the directory for future
'ensure' operations, and if you 'export' to a directory with stuff
already in it, it will only be partially overwritten (e.g. if you
export some package installing the file "A" and then export a package
installing file "B", you'll have both "A" and "B").

R=gbeaty, vadimsh

Bug: 1329641
Change-Id: I64253cee79de975ea58f805c63beef859b292a61
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3124202
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>