[milo][history] use page-specific test variant table implementation
In the past, this was done to reduce code duplication. However, the
tables in the test history page and the test results tab diverge
overtime, causing the test variant table implementation to be littered
with a lot of conditional statements. And this will get worse as we add
new features and implement different data loading mechanisms.
In this CL, we split out the implementation to two different sets of
components without changing the behavior of the pages (with the
exception that some GA tracking is removed). There are some code
duplication between the components but they will diverge further in
following CLs.

R=mdraz@google.com, nqmtuan@google.com

Change-Id: I215dc01bb58221870a0ba28a232f6bc30ecf385f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3670657
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>