[reflectutil] Add ShallowCopy.
This will allow safely generating shallow copies of proto messages, and
is an alternative to https://chromium-review.googlesource.com/c/3688745.

The downside of the protoc plugin approach is that it won't easily
allow importing this generated code to internal google repos, because
of stricter BUILD file rules on protos.

R=chanli, vadimsh

Change-Id: I7086135abbe05b3ec8b3eac0b0eaae3f552f3586
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3687321
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>