[environ] Hide Env map guts from the public API.
Them sticking out is a foot gun. One's natural assumption is that
the Env map maps an environment key to its value. But it is not.

One nasty side effect of this CL is that `environ.Env{}` still
compiles, but results in non-writable nil environ now. Not sure
how to catch this during compilation.

R=iannucci@chromium.org

Change-Id: I117c52b9bfecf46b2bc2977e3d59c5c892298ee2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3709843
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>