cv: handle cancel triggers completion event on more status
Reason: triggers could be cancelled when Run is PENDING as well. For
example, when Run is about to start, acl check finds that CL is not
approved hence decides to cancel the triggers and fail the run. Run
will be in PENDING status in that case. Right now, the completion
event is silently ignored leaving Run in PENDING status until a
CLUpdate event come in to bring Run to CANCELLED status because
LUCI CV thinks someone has removed the trigger on the CL.

R=robertocn

Change-Id: I5e281bbcecfa3652157da189f0acfd6fe08e1e28
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3738963
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>