Fix so that Python inherits file handles again on Windows.
The behavior was changed in go 1.17 so that only explicitly-specified
handles are inherited by subprocesses. Unfortunately, we have no
special knowledge of what handles the caller is attempting to pass to
the Python process, so we want to transparently pass everything
through.

This reverts to the previous behavior, mostly by copying the
1.16 implementation.

Bug: 1340653
Change-Id: I30302ffe2d661aa625e559385d5cfed7af731ef9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3739699
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Chenlin Fan <fancl@chromium.org>