[snoopy] Add a new end point for reporting pid.
PID is vaguely used to track a process' execution on a machine, this can
approximately determine when a recipe execution ends. It can be helpful
information to perform a cleanup actions or apply security policies.

Previously tasks used to self report their pid when starting, this
extends the opportunity to report a parent process id instead. This is
helpful when running in an isolated (jailed) environment where process
namespace can be different for inside vs outside.

Bug:1269830
Change-Id: I954922c0710f34c68d91227389a9634685a72aac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3741971
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>