cv: move all run notification to SaveTryjobs function
Currently, run is notified at the end which could be problematic if
the program crashes right after CV saves the tryjobs. This CL changes
the notification to be performed transactionally when saving the Tryjobs.
This may cause over-notification. But the cost is small

R=robertocn

Bug: 1312255
Change-Id: I34bc962c5e8ac4279472d9ba42b46001ea65b98a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3741189
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>