[gerrit] Add Comments field to ReviewInput struct
As part of the SetReview api, you can pass in a set of CommentInput
objects to create inline review comments. Attempts to use this
functionality are currently blocked due to the ReviewInput proto not
having the comments field defined, which causes any comments included as
input to get dropped in the Marshal.

Bug: fxb/104033
Change-Id: Ie8e72381c48e39e24debd60c1c77e7372e301975
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3764322
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Catherine Duncan <catduncan@google.com>