[realms] Do not store bindings for "uninteresting" permissions.
The only way to get realms.Permission value is to call
realms.RegisterPermission (*). And it is intended to be called at
init() time. Thus when we build structs that serve HasPermission
checks, we already know all possible permissions that will be
checked by the process and can discard all the rest to save RAM.

That optimization was the primary reason for introducing
RegisterPermission call. I punted on actually implementing it
though. But now realmset.Realms need a set of registered permissions
to build indexes for QueryRealm, so might just as well use this
opportunity to filter out "inactive" permissions.

*: There's also realms.GetPermissions, but it also returns only
previously registered permissions.

R=iannucci@chromium.org
BUG=1343842

Change-Id: Iac399df6e20f74cd5a9005e6c7a1a287d12e71e2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3767054
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>