[deploy] Refactoring in preparation for emitting notifications.
It turns out notifications are closely related to recording
historical log (we notify about a subset of recorded log entries),
so it makes sense to combine these portions of code into one
historyRecorder.

The only meh part is that we commit only finished AssetHistory
records (the history is immutable), but we want to send
ACTUATION_STARTING notification which is associated with ongoing
unfinalized AssetHistory entry. It necessitates having two
functions: recordAndNotify and notifyOnly.

Since some notifications depend on knowing how many actuations
failed in a row, keep this number in AssetHistory proto as well.

R=iannucci@chromium.org
BUG=1301023

Change-Id: Id9d310886af3e3321e68752daab7e850a0bc43a9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3759367
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>