[proto/protowalk] Add proto processing library.
This adds a protoreflect-based processing library which efficiently
allows processing of protobuf messages with field annotations applied.

This comes with processors which can check the `deprecated` proto
annotation as well as google.api.field_behavior = REQUIRED. It
additionally has a processor which can zero-out any
google.api.field_behavior = OUTPUT_ONLY fields.

This is meant to be used in conjunction with the protoc-gen-validate
generated code, though it would also be possible to implement a
processor which was able to apply these constraints via this
library instead, in case the generated code proves to be cumbersome.

R=chanli, ddoman, vadimsh

Bug: 1324799
Change-Id: I91837c9e5468d3e37346645f1e78259f7cf2774d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3632598
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>