[realms] Implement QueryRealms.
For all permissions marked with UsedInQueryRealms flag we keep
a list of all realms that have bindings for it grouped by project.
Then we walk over this list and check each individual binding.

This is not very advanced indexing structure, but we don't yet have
a lot of realms. And the function to check if a binding applies is
already well-optimized (in most cases it just compares a bunch of
integers). It should be fine for now.

This is only the "production" implementation. The implementation
for tests will be in a separate CL.

R=iannucci@chromium.org
BUG=1343842

Change-Id: I2ce40ae40f9f365535338001a05a9284fb31b8b2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3770947
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>