cv: process buildbucket pubsub message synchronously
Before, the approach is to send the message to a channel and spawn
several goroutines to read from the channel and batch processes
the message if the accumulated message reach batch size. This has
a problem that messages are acked and nacked asynchronously during
batch processing rather than in the callback when receiving the
message. This behaviour is actually discouraged because it may
create orphan message when messages is processed after `Receive`
function returns.
See: https://pkg.go.dev/cloud.google.com/go/pubsub#hdr-Receiving

The consequence is that the metric for Oldest unacked message age
is constantly high and is unalertable when anything goes wrong.
See: https://screenshot.googleplex.com/8csYNwru79XGXzw

This change switch the implementation to synchronous processing
the pubsub message. Although the code is simplified, the cost
is more load to datastore which is now effectively the same as
using push queue.

R=robertocn

Bug: 1257920
Change-Id: I328fa06637f5b20e14b22b3fac6b94fa5f013037
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3787319
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>