[led] Add support for editing CLs with https://<host>/<change> format.
This change updates the edit-cr-cl command to accept URLs that do not
specify the project. The format containing just the host and the change
is the format returned by git cl issue, which will enable mb try to use
the output of git cl issue to apply the CL corresponding to the local
git branch to the led job that it launches.

Bug: 1348269
Change-Id: I1fd8144051f87677646f93d599bf743df1e8f879
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3793005
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>