cv: make purgeReasons a new type that includes trigger information.
This is intended to allow separate purging of the different triggers
that may be present in a CL.

The new type distinguishes whether a CLError applies to a specific
trigger/set of triggers, or to all the active triggers.

To make this change safely revertible, it still writes to the deprecated
fields in the task and in the run state.

R=yiwzhang

Bug: 1242951
Change-Id: I3e4dcfd22d5ec1ec0ebe018d8e0a51bc66158b2a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3789728
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>