cv: fix CV can't find the corresponding execution if definition is unchanged
Before, we are relying on the pointer to find the match. However, if
a new config is ingested and we compute the new requirement, the pointer
to the definition in the new requirement may be different from the pointer
to the definition in the old requirement even if the definition has the
same content.

R=ddoman, robertocn

Bug: 1312255
Change-Id: I32d07da3668457dd22d1876f1436887e7cff1469
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3823152
Reviewed-by: Scott Lee <ddoman@chromium.org>
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>