[lucicfg][cv] rename QuotaPolicy to UserQuota
CV has a message, QuotaPolicy, and CV constructs
server/quota/proto:Policy from the configured QuotaPolicy. Because both
are named with "Policy", it produces unnecessary naming confusion, and
docs need to clarify which Policy it refers in a given context.

Originally, QuotaPolicy was designed to be generic enough such that
it could be used to represent quota policies for not only per-user, but
also for per-group, per-builder, and others.

However, it will likely that the message will be biased to the per-use
case further and further. If CV ever needs to support per-group, per-builder,
or any other than per-user, then it can add a new message optimized for
the particular grouping-level.

This simply renames QuotaPolicy to UserLimit. As of this moment, no LUCI
config sets values into cq_group.user_quotas, and this CL just renames
the messages and starlark parameters in a backward-incompatible way.

Bug: 1335519
Change-Id: Ie72d1e433befc97576870f98ab224e2377cc5a18
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3824005
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>